+++ NEWS.md
@@ -6,7 +6,7 @@ emmeans 1.4.5.009xxxxxxx
   * Improvements to `coxph` support for models with strata
   * `emmeans()` with `specs` of class `list` now passes any `offset` 
     and `trend` arguments (#179)
-    
+  * Added `plim` argument to `pwpp()` to allow controlling the scale
 
 emmeans 1.4.5
 -------------